Fix data race on i2c command contention
When multiple i2c commands are queued due to the lock, it was
theoretically possible for one command to be issued before the result of
the previous command was retrieved, resulting in data loss and a
communication error.

To fix the issue, the i2c lock is held not only throughout the issuing and
processing period of a command but also through the immediately following
data readout.

The code is refactored such that the i2c lock is used only in a single
function (sgp_read_from_cmd). As such most of sgp_i2c_read is moved there
and the remaining code is placed in sgp_verify_buffer() that performs the
CRC validation.