Return -EBUSY, not default values during iaq init
The upstream reviewers are not happy with the sensor returning default
values and wished for an -EBUSY error instead. This commit accomodates
for it.

Adjust iaq_init_skip_jiffies according to set init time on SGPC3,
also in any case default values are reported for 20s + init time (vs
~14.5s on the SGP30). Use 21s + init time to adjust for clock
imprecision.