Make IAQ thread the only accessor of measure_iaq
The IAQ thread now has exclusive access to the iaq measurement and has
the mandate to keep polling at the required rate.

The SGPC3's gas signals measurement is also exclusively handled
through the IAQ measurement

IAQ values are stored in a separate buffer that is not overwritten by
other communication with the sensor.

Version update to 0.6.0