Replace do_iaq_init with set_iaq_preheat_time
Replace the iaq-init interface that has a lasting effect (sets the
time for future iaq-inits) with one that only sets the init time
without actually triggering iaq-init.

Iaq init is still triggered automatically when setting the baseline or
changing the power mode.

set_iaq_preheat_time is only supported on the SGPC3

A value > 5min is considered a user error and may damage the hotplate