Remove permissions on unavailable sysfs attributes
Remove all permissions on sysfs attributes that are not available on
the plugged sensor.

While the kernel does provide a way to completely remove the
attribute instead of just removing permissions, it seems that the hook
(.is_visible) is not implemented on the IIO subsystem.

This patch also removes the linked getter and setter, in the case that
the root user, or a system daemon, manually changes the permissions
back. This scenario then result in an I/O Error being reported.